pyproject.toml
[project]
name = "discord-bot"
version = "2.1.0-beta.1"
description = "Yet Another Shitly Made Discord Bot."
license = "AGPL-3.0-or-later"
readme = "docs/README.md"
requires-python = ">=3.14"
dependencies = [
	"discord-py>=2.6.4",
    "discord-py[voice]>=2.6.4",
    "google-genai>=1.62.0",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "yt-dlp>=2026.2.4",
]


[tool.basedpyright]
typeCheckingMode = "standard"
reportOptionalMemberAccess = false
reportFunctionMemberAccess = false

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.0",
]

[project.scripts]
bot = "main:main"

[tool.uv]
package = true

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.hatch.build.targets.wheel.sources]
"src" = ""

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

## --- RUFF ---
[tool.ruff]
select = ["ALL"]
# Ruff configuration for PEP8 compliance
# See https://beta.ruff.rs/docs/settings/ for more information
target-version = "py314" #  3.14159265359!
extend-exclude = [
    "*.py.example",
]

[tool.ruff.lint]
# Enable the following rule categories:
# E: pycodestyle errors
# F: Pyflakes
# I: isort
# N: pep8-naming
# W: pycodestyle warnings
# B: flake8-bugbear
# C: flake8-comprehensions
# D: pydocstyle
# ANN: flake8-annotations
# PTH: flake8-use-pathlib
select = [
    "E", "F", "I", "N", "W", "B", "C", "D", "ANN", "PTH", "C90",
]

# Ignore specific rules if needed
ignore = [
    "D100", # Missing docstring in public module (often redundant)
    "D104", # Missing docstring in public package
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.mccabe]
max-complexity = 12
